Feat: 코드 리뷰 수정 사항 반영
Ref: #5
